# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Precifica3D API",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/v1/history": {
      "post": {
        "tags": [
          "history"
        ],
        "summary": "Record Snapshot",
        "description": "Record a snapshot — the ONLY writer of frozen fields, and idempotent by construction.\n\n`ON CONFLICT DO NOTHING` + read-back is what makes a retry an idempotent SUCCESS rather than a\nduplicate or an error: the outbox may replay the same `clientSnapshotId` any number of times\n(lost response, reconnect, app restart, two tabs) and always gets back the row the server\nalready created. Because the unique key is UNCONDITIONAL — tombstones included — a replay that\narrives AFTER the seller deleted the entry cannot RESURRECT it: the insert no-ops, and the\nread-back finds only the soft-deleted row, which is not served.",
        "operationId": "record_snapshot_api_v1_history_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnapshotIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotOut"
                }
              }
            }
          },
          "200": {
            "description": "Already recorded — the same clientSnapshotId was replayed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Persistence requires an active premium grant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request body (unparseable)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Per-field validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "history"
        ],
        "summary": "List History",
        "description": "Newest-first by the DEVICE's date — the date that is the seller's claim (FR-523).\n\nKeyset pagination (data-model D4), never OFFSET and never a silent cap: a page size is NOT a\nlimit on how many snapshots a seller may keep (that would be a business-rules amendment). The\nclient follows ``nextCursor`` to exhaustion — or, from PR-B, lazily via [Carregar mais] — so the\nwhole unbounded history is reachable. ``limit + 1`` peeks one row ahead to decide whether a next\npage exists.\n\nUS6 filters (ux §2.6), all owner-scoped and composing with the keyset:\n\n* ``q`` — a substring label search, ILIKE (case-insensitive) and **accent-sensitive** (F5,\n  owner-accepted for E4: ``joao`` does not find ``João``). ILIKE metacharacters in the term are\n  escaped, so a literal ``%``/``_`` in a label is matched, not treated as a wildcard. A NULL\n  label never matches (an unlabelled row is not a search hit).\n* ``from``/``to`` — a range on the **DEVICE** date (the seller's claimed quote date), never\n  ``created_at`` (which is unverifiable metadata and never ordered/filtered by, FR-528).\n* ``clientSnapshotId`` — an exact lookup, so the detail can resolve a deep-linked snapshot that\n  may not be on the first lazily-loaded page (review PR-A M3, preserved under lazy pagination).",
        "operationId": "list_history_api_v1_history_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "clientSnapshotId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Clientsnapshotid"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotPage"
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Persistence requires an active premium grant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request body (unparseable)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Per-field validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope"
      },
...
      "SnapshotIn": {
        "properties": {
          "clientSnapshotId": {
            "type": "string",
            "format": "uuid",
            "title": "Clientsnapshotid"
          },
          "kind": {
            "type": "string",
            "enum": [
              "SINGLE",
              "KIT"
            ],
            "title": "Kind"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "quoteValidityDays": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quotevaliditydays"
          },
          "deviceQuotedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Devicequotedat"
          },
          "deviceUtcOffsetMinutes": {
            "type": "integer",
            "title": "Deviceutcoffsetminutes"
          },
          "modelVersion": {
            "type": "string",
            "title": "Modelversion"
          },
          "headlineTotal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Headlinetotal"
          },
          "headlineBasis": {
            "type": "string",
            "enum": [
              "PRECO_VAREJO",
              "PRECO_ATACADO"
            ],
            "title": "Headlinebasis"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "clientSnapshotId",
          "kind",
          "deviceQuotedAt",
          "deviceUtcOffsetMinutes",
          "modelVersion",
          "headlineTotal",
          "headlineBasis",
          "payload"
        ],
        "title": "SnapshotIn",
        "description": "A recorded claim. Self-contained and frozen on the device at record time — the outbox stores\nexactly this body and replays it byte-for-byte."
      },
...
      "SnapshotOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "clientSnapshotId": {
            "type": "string",
            "format": "uuid",
            "title": "Clientsnapshotid"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "quoteValidityDays": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quotevaliditydays"
          },
          "deviceQuotedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Devicequotedat"
          },
          "deviceUtcOffsetMinutes": {
            "type": "integer",
            "title": "Deviceutcoffsetminutes"
          },
          "modelVersion": {
            "type": "string",
            "title": "Modelversion"
          },
          "payloadSchemaVersion": {
            "type": "integer",
            "title": "Payloadschemaversion"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "headlineTotal": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Headlinetotal"
          },
          "headlineBasis": {
            "type": "string",
            "title": "Headlinebasis"
          }
        },
        "type": "object",
        "required": [
          "id",
          "clientSnapshotId",
          "kind",
          "label",
          "quoteValidityDays",
          "deviceQuotedAt",
          "deviceUtcOffsetMinutes",
          "modelVersion",
          "payloadSchemaVersion",
          "payload",
          "headlineTotal",
          "headlineBasis"
        ],
        "title": "SnapshotOut"
      },
      "SnapshotPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SnapshotOut"
            },
            "type": "array",
            "title": "Items"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "SnapshotPage"
      },
...
    }
  }
}